+++ CHANGELOG.md
@@ -5,9 +5,24 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [7.0.7] - 2025-09-22
+
+### Security
+
+- Bumped `axios` to 1.12.2 to mitigate [CVE-2025-58754](https://avd.aquasec.com/nvd/cve-2025-58754)
+
+### Changed
+
+- Modified sourcebucketpattern to allow valid s3 bucket names
+
+### Removed
+
+- AppRegistry application at resource level
+
 ## [7.0.6] - 2025-07-28
 
 ### Security
+
 - Bump `form-data` to mitigate [CVE-2025-7783](https://github.com/advisories/GHSA-fjxv-7rqg-78g4)
 
 ## [7.0.5] - 2025-07-07